Reorganize README
This is an attempt to make the readme more concise.

I added a few new sections and moved the install section and the api
usage section to their own files.

A build tag, description, contributing, and contributor section were
added. There could probably be one or two more short sections to add to
make the readme more complete, but I think this is a good start to make
the readme more concise and easier to navigate.