Create base interface test
Created unit tests for the base model. This was implemented in a
separate test file so that it can be used to test inherited classes
confirm to the interface of the base model.

Created a new method find. This is done to replace the find_by_id
method, because the find method will automatically return a 404 status
code and custom json message. This will reduce the number of
conditionals needed in the endpoints and the number of integration tests
needed.