Remove init method
Removed init method as SQLAlchemy adds an implicit constructor to all
model classes which accepts keyword arguments for all its columns and
relationships. If we need to add custom stuff to an init method then we
can add it back in at that time, and call super to keep accepting kwargs.