Remove unused dependencies (#105)
* Remove unused dependencies

There were many dependencies installed back in march that were never
used. One dep, flask-script, was actually archived and unmaintained.
Removed the duplicated freezegun dependency. All other deps removed were
simply not being used. I did leave 2 deps in the Pipfile that we are
currently not using, deps such as the postgres adapter, and jinja.

* Remove sqlalchemy dep

Flask-sqlalchemy already requires this as a dep. We should allow
Flask-sqlalchemy to install the most updated version it can run with.