Add stubs to base_model db methods
Added stubs to the base_model db methods. This enabled me to remove the
valid payload dependency on all the endpoint tests. A valid payload for
the lease resource is now independently tested in the schema test file.

These tests validate that we're interacting with the model layer
correctly and returning a proper response.