Remove user init method test
This test doesn't really test anything that isn't tested in numerous
places (the ability to create a user). It has a comment highlighting the
fact that there is a dependency on the order of the arguments. This is
an override of the default kwarg behavior provided by flask-restful. All
init methods will be removed unless we need to have custom behavior
which we will then preserve the kwarg behavior by calling super.