Speeding up a slow test (#177)
* Speed up the UserInvite integration test

Sped up the test by mock-asserting the send_msg_email is sent. The test
was slow because it was waiting for a response from the Email client to
know if the statement executed successfully or not. Ideally sending
emails should be done in the background and mocked automatically in the
tests (maybe someday). However, asserting the message is sent would need
to be done regardless.

* Remove archived argument

Passing an archived value is not necessary unless we want to specify
that it should be True. By default it will be None in the column which
results in Falsey.

* Default archive to False

* Remove unused import