Create test for validating propertyManagerIDs
A section of the validation code was untested. This adds a test for the
missing propertyManagerIDs validation, and updates the statement to use
load instead of the validate method. This will provide a more complete
error response message.

The many=True was removed from the validate method as it is no longer
needed.