Exclude __init__.py files from Flake8 for now
Some of the __init__.py files are used for imports.
Flake8 raises an unused import error for these and also complains
if the import was a * import.

The unused import error can be resolved by putting the name of the
class or function being imported into __all__, but there will still
be an error if a * import is used.

Since we are still actively creating fixtures and other tests that
might be imported through these __init__ files, I think we should
leave the files be for now and just have Flake8 skip over them during
the pre-commit check.