Change create_lease into func call for some tests
In these tests:
- `test_pm_is_authorized_to_get_all`
- `test_staff_are_authorized_to_get_all`
- `test_admin_is_authorized_to_get_all`

creat_lease() was assigned to the variable `lease`, but the variable
was never used.  I couldn't tell if this function call was needed
(the tests still passed when the line was removed), but the variable
name definitely wasn't.

I just changed the line to a bare function call instead of a variable
assignment in order to avoid an unused variable warning.