Remove PropertyUpdateSchema, pass context instead
Instead of using a new schema, this solution passes a context dictionary
to the PropertySchema instead.  Right now, the context dictionary is
only passed in from the Property resource PUT function.

When the `update` function is called with this context, the
PropertySchema is able to access the information from its validation
methods and handle the validation using conditionals based on what's in
the context dictionary.

In order for this to work, an optional parameter had to be added to the
BaseModel's `update` and `validate` methods and a context dictionary has
to be added manually for certain schema tests (the ones testing the PUT
functionality).