squash a bug in returning the tenants list for users with role == property manager
The tricky part is the tenants() method on the property model returns an array of json. When iterating over all properties the list gains a second dimension.
For usability to the front end, this 2D list needs to be flattened. After some StackOverflow reading, I decided to use chain from the itertools library for array flattening.