Remove find_by_id method (#243)
Changed all calls to find_by_id to the find method. Removed conditionals
that checks if the find_by_id returned successfully as they're not
needed with the find method since it will automatically return a 404
with an appropriate response. Removed any test that was failing because
it was asserting incorrect behavior. One test is now skipped, because
the current logic seems incorrect, and the comments also support this.