Initial commit for moving toward MethodViews
This switches the tenant endpoint to use MethodViews instead of the now,
or soon to be, deprecated Flask-RESTful library. The only benefit we were
getting from using flask-restful was that all our responses were configured
to be json responses where as by default flask responds with text/html
when given a string.

This caused some issues where we call abort, to raise an error, because
those responses are automatically structured to return text instead of
a json response. So to get around this we will not use the abort method
in our app and instead use our own error response class that will return
an appropriate json response.