Remove tenant endpont DELETE action
Removed the `delete()` method from this resource and refactored the
update tenant endpoint test.

The `delete()` method was being used for archiving tenants through a
"toggle" mechanism that would archive the tenant if they were active and
unarchive them if they were already archived.

This required the front end to be aware of a tenant's archival status
before making the request, or else risk accidentally doing the opposite
of what was intended.

Rather than risk this and write additional logic for custom
messages on the back end, we decided to switch to a more deterministic
(and simple) approach by using the existing PUT action to handle
archival.

Success messages can be handled by the front end (as with all update
actions) and the built-in error messages are returned by the back end.