Do not revoke admin jwt when they're archiving users
The `get_jwt` method returns the jwt for the person who accessed the
endpoint. The logic in the archive user would revoke the admin, not the
user that is being archived. In order to prevent an archived user from
accessing the site we must check user credentials at auth instead of
relying on the jwt data. Ticket codeforpdx/dwellingly-app/issues/660 has
been created to fix this.