Add DummyResource and Gatekeeper tests
These tests hit a dummy endpoint that utilizes the `allowed_params`
decorator to limit what json fields are passed to the endpoint. The
dummy endpoint is only included in the "testing" environment.

One additional thing to note is that if you alter the dummy resource's
set of allowed parameters, it will persist through the remainder of your
tests.  If this is not desired, the `.clear()` method should be used on
the set.