Add gatekeeper utility to check request params (#321)
* Add gatekeeper utility to check request params

Use the `allowed_params` decorator to restrict an endpoint to only
accept certain fields in the JSON request.

Decorator takes a set of strings as its only argument.

* Update which envrionments trigger GatekeeperError

* Add DummyResource and Gatekeeper tests

These tests hit a dummy endpoint that utilizes the `allowed_params`
decorator to limit what json fields are passed to the endpoint. The
dummy endpoint is only included in the "testing" environment.

One additional thing to note is that if you alter the dummy resource's
set of allowed parameters, it will persist through the remainder of your
tests.  If this is not desired, the `.clear()` method should be used on
the set.

* Remove Dummy Resource from routes.py

Rather than adding in a new route just for testing, updated tests to
use flask.current_app.test_request_context.

* Remove `request` import and fix `put()` call