Do not import seed file at top level of manage.py (#373)
* Do not import seed file at top level of manage.py

Importing the seed file at the top level causes all environments to
parse the seed file. Which causes issues for the production environment,
because the seed file contains non-production libraries that are
imported. Plus we don't want to import/run any code that is not
necessary in Production.

* Flask can't find these models unless imported here

Since the seed file is no longer parsed (which had initially imported a
ton of the application files) we are receiving errors due to missing
models. Looks like Flask can't load them unless explicitly told to so here.