Fix tenant create endpoint
The endpoint was adding empty strings to the lease param even if the
lease param was not sent. Updated to only include the params the
frontend sends.

This is needed so that only the params that are provided are validated;
otherwise the endoint was adding empty lease params when the lease is not
needed will cause the tenant not to be created due to invalid lease params.

Fixes codeforpdx/dwellingly-app/issues/692