Allow api to accept the actual values of a Ticket
The previous implementation created a hard dependency when working with
a Ticket. The requestor needed to know the keys used in the backed and
how those keys mapped to the actual ticket status.

Because of this when someone made a request and wanted to move the
ticket to `In Progress` they would need to know that they had to send
the string `In_Progress`.

I decided to remove the enum that was in use, and changed the column to
a String. The TicketStatus was not being used in logic anywhere, and so
it seemed its only purpose was to validate the strings that went into
the db. To make things a bit more simpler it is removed for now, and a
straight tuple is created that is used to validate what goes into the
db.