Switch to has_one relationship for tenant/lease
Updates to use a has_one relationship. Tenant has one lease.

This is done by defining the tenant_id field on the Lease table
to be unique so that only one lease record can be inserted for a tenant,
and by changing the relationship not to return a collection object.