Initial work to remove role column from UserModel (#358)
* Initial work to remove role column from UserModel

We don't really have different roles so they are being removed. We have
different types of users which have been implemented using STI.

This is the first phase to remove roles from the backend. We will still
have a role per se as far as the frontend code is concerned.

The next phase will be to update the frontend to move off of integers as
their Role, and use the typed strings. Along with changing any hard
coded dependencies on Roles. The final phase will be back in the backend
to remove the final remnants of Roles.

Doing it in a 3 phase step as to not break anything for other
developers :-)

* Refactor UserLogin

* We still need to assign a role on update

Still need to assign a role until the frontend is updated not to use
roles.

* User singular name for user_type class