Overmind: Config to start and end everything
Install tmux and overmind and copy the `.overmind.env.example` file to
`overmind.env` and copy the `Procfile.dev.example` to `Procfile.dev`
then adjust the location of your frontend repo in the Procfile. Default
assumes that the frontend repo is located in the same folder as the
backend repo. For ex: Change `cd ../dwellingly-app` to `cd
../your_relative_location` or use an absolute path.

Start all servers with `overmind start`. You can connect to any server
in the terminal with `overmind connect name_of_process`.
for ex: `overmind connect frontend` or `overmind connect backend`

This will start a development frontend and backend server. And start a
testing backend and frontend server.

Simplifies start up and removes the need to set config to specify which
server to start.